requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
absl-py==1.4.0
altair==4.2.2
apturl==0.5.2
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
binaryornot==0.4.4
blinker==1.4
boto3==1.26.103
botocore==1.29.103
Brlapi==0.8.3
cachetools==5.0.0
certifi==2022.12.7
chardet==4.0.0
click==8.0.4
cmake==3.26.1
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
contourpy==1.0.7
cookiecutter==2.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
executing==1.2.0
fasteners==0.14.1
filelock==3.10.7
fire==0.5.0
fonttools==4.39.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-api-python-client==1.7.11
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.53.0
httplib2==0.20.2
huggingface-hub==0.13.3
idna==2.10
importlib-metadata==4.6.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybboxes==0.1.6
pycairo==1.20.1
pycups==2.0.1
pydeck==0.8.0
PyDrive==1.3.1
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.3
roboflow==1.0.2
rsa==4.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sahi==0.11.13
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.1
semver==2.13.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.31
stack-data==0.6.2
streamlit==1.20.0
sympy==1.11.1
systemd-python==234
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminaltables==3.1.10
text-unidecode==1.3
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.3
uritemplate==3.0.1
urllib3==1.26.15
usb-creator==0.3.7
validators==0.20.0
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wget==3.2
xdg==5
xkit==0.0.0
yolov5==7.0.11
zipp==1.0.0
